requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
openai==1.3.0
sentence-transformers==2.2.2
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
aiopg==1.4.0
ollama==0.1.6
httpx==0.25.2
python-dotenv==1.0.0
numpy==1.24.3
xmltodict==0.13.0
python-multipart==0.0.6
pgvector==0.2.0
torch==2.1.0
transformers==4.35.2
